+++ src/entities/trade-post/model/mutations.ts
@@ -1,8 +1,18 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { deleteTradePostLike, postTradePostLike } from '@/entities/trade-post/api/apis';
+import {
+  deleteTradePost,
+  deleteTradePostLike,
+  patchTradePost,
+  postTradePostLike,
+} from '@/entities/trade-post/api/apis';
 
-import type { AllPost } from '@/entities/trade-post/lib/types';
+import type {
+  AllPost,
+  DeletePostResponse,
+  UpdatePostRequest,
+  UpdatePostResponse,
+} from '@/entities/trade-post/lib/types';
 
 export const usePostTradePostLikeMutation = () => {
   const queryClient = useQueryClient();
@@ -53,6 +63,30 @@ export const useDeleteTradePostLikeMutation = () => {
         queryClient.setQueryData(['trade-posts'], context.previousPosts);
       }
       console.error('좋아요 취소 실패', error);
+    },
+  });
+};
+
+export const useDeleteTradePostMutation = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation<DeletePostResponse, Error, number>({
+    mutationFn: deleteTradePost,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['trade-posts'] });
+      queryClient.invalidateQueries({ queryKey: ['trade', 'detail'] });
+    },
+  });
+};
+
+export const useUpdateTradePostMutation = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation<UpdatePostResponse, Error, { postId: number; data: UpdatePostRequest }>({
+    mutationFn: ({ postId, data }) => patchTradePost(postId, data),
+    onSuccess: (data, { postId }) => {
+      queryClient.invalidateQueries({ queryKey: ['trade-posts'] });
+      queryClient.invalidateQueries({ queryKey: ['trade', 'detail', postId] });
     },
   });
 };
